INT-269 upgrade dependencies and split source and tests
