+++ .vscodeignore
@@ -3,6 +3,7 @@
 out/**
 node_modules/**
 src/**
+test/**
 .gitignore
 .yarnrc
 webpack.config.js
