+++ package.json
@@ -10,7 +10,7 @@
 		"url": "https://github.com/intuita-inc/intuita-vscode-extension"
 	},
 	"engines": {
-		"vscode": "^1.72.0"
+		"vscode": "^1.73.0"
 	},
 	"categories": [
 		"Programming Languages",
@@ -264,22 +264,21 @@
 		"compile": "webpack",
 		"watch": "webpack --watch",
 		"package": "webpack --mode production --devtool hidden-source-map",
-		"lint:eslint": "eslint src --fix --ext ts",
-		"lint:prettier": "prettier --write src",
+		"lint:eslint": "eslint src test --fix --ext ts",
+		"lint:prettier": "prettier --write src test",
 		"test": "mocha"
 	},
 	"devDependencies": {
-		"@types/chai": "^4.3.1",
-		"@types/form-data": "^2.5.0",
+		"@types/chai": "^4.3.4",
 		"@types/glob": "^8.0.0",
 		"@types/mocha": "^10.0.0",
-		"@types/node": "18.11.2",
-		"@types/vscode": "^1.72.0",
-		"@typescript-eslint/eslint-plugin": "^5.40.1",
-		"@typescript-eslint/parser": "^5.40.1",
-		"@vscode/test-electron": "^2.1.5",
-		"chai": "^4.3.6",
-		"eslint": "^8.25.0",
+		"@types/node": "18.11.9",
+		"@types/vscode": "^1.73.1",
+		"@typescript-eslint/eslint-plugin": "^5.43.0",
+		"@typescript-eslint/parser": "^5.43.0",
+		"@vscode/test-electron": "^2.2.0",
+		"chai": "^4.3.7",
+		"eslint": "^8.27.0",
 		"eslint-config-prettier": "^8.5.0",
 		"eslint-plugin-prettier": "^4.2.1",
 		"glob": "^8.0.3",
@@ -288,7 +287,7 @@
 		"ts-loader": "^9.4.1",
 		"ts-node": "^10.9.1",
 		"umd-compat-loader": "^2.1.2",
-		"webpack": "^5.72.1",
+		"webpack": "^5.75.0",
 		"webpack-cli": "^4.10.0"
 	},
 	"dependencies": {
