+++ test/buildReplacement.test.ts
@@ -1,9 +1,7 @@
 import { assert } from 'chai';
-import {
-	buildReplacement,
-	buildTs2769ObjectAssignReplacement,
-} from '../features/repairCode/buildReplacement';
+
 import { factory } from 'typescript';
+import { buildReplacement, buildTs2769ObjectAssignReplacement } from '../src/features/repairCode/buildReplacement';
 
 describe('buildReplacement', () => {
 	it("should change the '' string into the 0 number", () => {
