+++ test/entry.ts
@@ -1,0 +1,8 @@
+export type Entry = Readonly<{
+    startLineNumber: number,
+    startColumn: number,
+    endLineNumber: number,
+    endColumn: number,
+    code: number,
+    message: string,
+}>;
